Rename internal/utils to internal/netutil
The three files form one cohesive unit: the resilient HTTP client composed
from the fast-timeout retry transport and the hf-mirror workaround. Name the
package for what it is instead of a grab-bag utils.
