+++ internal/netutil/fast_timeout_retry.go
@@ -1,4 +1,4 @@
-package utils
+package netutil
 
 import (
 	"context"
