+++ internal/netutil/fix_hf_mirror.go
@@ -1,4 +1,4 @@
-package utils
+package netutil
 
 import (
 	"net/http"
