+++ internal/netutil/httpclient.go
@@ -1,4 +1,4 @@
-package utils
+package netutil
 
 import (
 	"context"
